# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PteroStore Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (user's own or all for admin)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/payment-proof:
    post:
      operationId: uploadPaymentProof
      tags: [payments]
      summary: Upload payment proof (base64)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Discounts
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}/reply:
    post:
      operationId: replyReview
      tags: [reviews]
      summary: Admin reply to review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewReplyInput"
      responses:
        "200":
          description: Replied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Bug Reports
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
  /settings/faqs:
    get:
      operationId: listFaqs
      tags: [settings]
      summary: List FAQs
      responses:
        "200":
          description: List of FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"

    post:
      operationId: createFaq
      tags: [settings]
      summary: Create FAQ (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /settings/faqs/{id}:
    patch:
      operationId: updateFaq
      tags: [settings]
      summary: Update FAQ (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"

    delete:
      operationId: deleteFaq
      tags: [settings]
      summary: Delete FAQ (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        totalOrders:
          type: ["integer", "null"]
        totalSpent:
          type: ["number", "null"]
      required:
        - id
        - username
        - email
        - role
        - createdAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        description:
          type: string
        detail:
          type: string
        usageInfo:
          type: string
        suitableFor:
          type: string
        benefits:
          type: string
        badge:
          type: ["string", "null"]
          description: "popular, recommended, new, hot"
        status:
          type: string
          enum: [ready, sold_out, coming_soon]
        stock:
          type: ["integer", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        eligibleForInviteDiscount:
          type: boolean
      required:
        - id
        - name
        - slug
        - category
        - price
        - description
        - detail
        - usageInfo
        - suitableFor
        - benefits
        - status
        - isActive
        - sortOrder
        - createdAt
        - eligibleForInviteDiscount
    ProductInput:
      type: object
      required: [name, category, price, description, detail, usageInfo, suitableFor, benefits]
      properties:
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        description:
          type: string
        detail:
          type: string
        usageInfo:
          type: string
        suitableFor:
          type: string
        benefits:
          type: string
        badge:
          type: ["string", "null"]
        status:
          type: string
          enum: [ready, sold_out, coming_soon]
        stock:
          type: ["integer", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        eligibleForInviteDiscount:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        description:
          type: string
        detail:
          type: string
        usageInfo:
          type: string
        suitableFor:
          type: string
        benefits:
          type: string
        badge:
          type: ["string", "null"]
        status:
          type: string
          enum: [ready, sold_out, coming_soon]
        stock:
          type: ["integer", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        eligibleForInviteDiscount:
          type: boolean
    Order:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        userId:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        user:
          $ref: "#/components/schemas/User"
        originalPrice:
          type: number
        discountAmount:
          type: number
        finalPrice:
          type: number
        discountCode:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [dana, qris]
        paymentProofUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, paid, confirmed, failed, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - userId
        - productId
        - originalPrice
        - discountAmount
        - finalPrice
        - paymentMethod
        - status
        - createdAt
        - updatedAt
    OrderInput:
      type: object
      required: [productId, paymentMethod]
      properties:
        productId:
          type: integer
        paymentMethod:
          type: string
          enum: [dana, qris]
        discountCode:
          type: string
        notes:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, paid, confirmed, failed, cancelled]
        notes:
          type: string
    PaymentProofInput:
      type: object
      required: [proofBase64]
      properties:
        proofBase64:
          type: string
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: ["integer", "null"]
        user:
          $ref: "#/components/schemas/User"
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        adminReply:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - rating
        - comment
        - createdAt
    ReviewInput:
      type: object
      required: [rating, comment]
      properties:
        productId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    ReviewReplyInput:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
...
    SiteSettings:
      type: object
      properties:
        id:
          type: integer
        siteName:
          type: string
        siteDescription:
          type: string
        whatsappNumber:
          type: string
        danaNumber:
          type: string
        danaName:
          type: string
        qrisImageUrl:
          type: ["string", "null"]
        maintenanceMode:
          type: boolean
        maintenanceMessage:
          type: ["string", "null"]
        bannerEnabled:
          type: boolean
        bannerText:
          type: ["string", "null"]
        bannerColor:
          type: ["string", "null"]
      required:
        - id
        - siteName
        - siteDescription
        - whatsappNumber
        - danaNumber
        - danaName
        - maintenanceMode
        - bannerEnabled
    SiteSettingsUpdate:
      type: object
      properties:
        siteName:
          type: string
        siteDescription:
          type: string
        whatsappNumber:
          type: string
        danaNumber:
          type: string
        danaName:
          type: string
        qrisImageUrl:
          type: ["string", "null"]
        maintenanceMode:
          type: boolean
        maintenanceMessage:
          type: ["string", "null"]
        bannerEnabled:
          type: boolean
        bannerText:
          type: ["string", "null"]
        bannerColor:
          type: ["string", "null"]
    Faq:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        sortOrder:
          type: integer
      required:
        - id
        - question
        - answer
        - sortOrder
    FaqInput:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
        sortOrder:
          type: integer
    FaqUpdate:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        sortOrder:
          type: integer