# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PteroStore Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputStatus } from './productInputStatus';

export interface ProductInput {
  name: string;
  slug?: string;
  category: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  description: string;
  detail: string;
  usageInfo: string;
  suitableFor: string;
  benefits: string;
  /** @nullable */
  badge?: string | null;
  status?: ProductInputStatus;
  /** @nullable */
  stock?: number | null;
  isActive?: boolean;
  sortOrder?: number;
  eligibleForInviteDiscount?: boolean;
}