# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PteroStore Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  id: number;
  siteName: string;
  siteDescription: string;
  whatsappNumber: string;
  danaNumber: string;
  danaName: string;
  /** @nullable */
  qrisImageUrl?: string | null;
  maintenanceMode: boolean;
  /** @nullable */
  maintenanceMessage?: string | null;
  bannerEnabled: boolean;
  /** @nullable */
  bannerText?: string | null;
  /** @nullable */
  bannerColor?: string | null;
}